requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf>=4.25.0
